backend/requirements.txt
Flask
flask-cors
flask-mysqldb
python-dotenv
werkzeug
Flask-JWT-Extended

